add new scouter
