+++ py-opsml/python/opsml/model/__init__.pyi
@@ -1,15 +1,15 @@
 from pathlib import Path
-from typing import Any, Dict, Optional, List
+from typing import Any, Dict, List, Optional
 
 from ..core import CommonKwargs, FeatureSchema, OnnxSchema
 from ..data import DataType
 from ..scouter import (
-    SpcDriftProfile,
-    PsiDriftProfile,
     CustomDriftProfile,
-    SpcDriftConfig,
-    PsiDriftConfig,
     CustomMetricDriftConfig,
+    PsiDriftConfig,
+    PsiDriftProfile,
+    SpcDriftConfig,
+    SpcDriftProfile,
 )
 
 class HuggingFaceORTModel:
