+++ py-opsml/python/opsml/scouter/__init__.py
@@ -1,106 +1,19 @@
+# pylint: disable=no-name-in-module
 # type: ignore
 
-from .. import scouter
+from .scouter import (  # noqa
+    alert,
+    client,
+    drift,
+    logging,
+    observe,
+    profile,
+    queue,
+    types,
+)
+from .version import __version__
 
-DriftProfile = scouter.DriftProfile
-DriftType = scouter.DriftType
-SpcServerRecord = scouter.SpcServerRecord
-ServerRecord = scouter.ServerRecord
-ServerRecords = scouter.ServerRecords
-CommonCrons = scouter.CommonCrons
-AlertZone = scouter.AlertZone
-SpcAlertType = scouter.SpcAlertType
-SpcAlertRule = scouter.SpcAlertRule
-AlertDispatchType = scouter.AlertDispatchType
-PsiAlertConfig = scouter.PsiAlertConfig
-SpcAlertConfig = scouter.SpcAlertConfig
-SpcFeatureDriftProfile = scouter.SpcFeatureDriftProfile
-PsiDriftConfig = scouter.PsiDriftConfig
-SpcDriftConfig = scouter.SpcDriftConfig
-SpcDriftProfile = scouter.SpcDriftProfile
-Distinct = scouter.Distinct
-Quantiles = scouter.Quantiles
-Histogram = scouter.Histogram
-NumericStats = scouter.NumericStats
-CharStats = scouter.CharStats
-WordStats = scouter.WordStats
-StringStats = scouter.StringStats
-FeatureProfile = scouter.FeatureProfile
-DataProfile = scouter.DataProfile
-FeatureDrift = scouter.FeatureDrift
-SpcFeatureDrift = scouter.SpcFeatureDrift
-SpcDriftMap = scouter.SpcDriftMap
-PsiDriftProfile = scouter.PsiDriftProfile
-Feature = scouter.Feature
-Features = scouter.Features
-PsiFeatureQueue = scouter.PsiFeatureQueue
-SpcFeatureQueue = scouter.SpcFeatureQueue
-Bin = scouter.Bin
-PsiFeatureDriftProfile = scouter.PsiFeatureDriftProfile
-PsiDriftMap = scouter.PsiDriftMap
-PsiServerRecord = scouter.PsiServerRecord
-AlertThreshold = scouter.AlertThreshold
-CustomMetricAlertCondition = scouter.CustomMetricAlertCondition
-CustomMetricAlertConfig = scouter.CustomMetricAlertConfig
-CustomMetricDriftConfig = scouter.CustomMetricDriftConfig
-CustomMetric = scouter.CustomMetric
-CustomDriftProfile = scouter.CustomDriftProfile
-CustomMetricServerRecord = scouter.CustomMetricServerRecord
-DataProfiler = scouter.DataProfiler
-DataType = scouter.DataType
-ScouterError = scouter.ScouterError
-Drifter = scouter.Drifter
-DriftProfile = scouter.DriftProfile
+Drifter = drift.Drifter
+DataProfiler = profile.DataProfiler
 
-
-__all__ = [
-    "DriftProfile",
-    "DriftType",
-    "SpcServerRecord",
-    "ServerRecord",
-    "ServerRecords",
-    "CommonCrons",
-    "AlertZone",
-    "SpcAlertType",
-    "SpcAlertRule",
-    "AlertDispatchType",
-    "PsiAlertConfig",
-    "SpcAlertConfig",
-    "SpcFeatureDriftProfile",
-    "PsiDriftConfig",
-    "SpcDriftConfig",
-    "SpcDriftProfile",
-    "Distinct",
-    "Quantiles",
-    "Histogram",
-    "NumericStats",
-    "CharStats",
-    "WordStats",
-    "StringStats",
-    "FeatureProfile",
-    "DataProfile",
-    "FeatureDrift",
-    "SpcFeatureDrift",
-    "SpcDriftMap",
-    "PsiDriftProfile",
-    "Feature",
-    "Features",
-    "PsiFeatureQueue",
-    "SpcFeatureQueue",
-    "Bin",
-    "PsiFeatureDriftProfile",
-    "PsiDriftMap",
-    "PsiServerRecord",
-    "AlertThreshold",
-    "CustomMetricAlertCondition",
-    "CustomMetricAlertConfig",
-    "CustomMetricDriftConfig",
-    "CustomMetric",
-    "CustomDriftProfile",
-    "CustomMetricServerRecord",
-    "DataProfiler",
-    "DataType",
-    "ScouterError",
-    "Drifter",
-    "DriftProfile",
-]
+__all__ = ["__version__", "Drifter", "DataProfiler"]
