+++ py-opsml/python/opsml/scouter/logging/__init__.py
@@ -1,0 +1,15 @@
+# type: ignore
+
+from .. import logging  # noqa: F401
+
+LogLevel = logging.LogLevel
+WriteLevel = logging.WriteLevel
+LoggingConfig = logging.LoggingConfig
+RustyLogger = logging.RustyLogger
+
+__all__ = [
+    "LogLevel",
+    "WriteLevel",
+    "LoggingConfig",
+    "RustyLogger",
+]
