+++ py-opsml/python/opsml/scouter/profile/__init__.py
@@ -1,0 +1,28 @@
+# type: ignore
+
+from .. import profile
+
+CharStats = profile.CharStats
+DataProfile = profile.DataProfile
+DataProfiler = profile.DataProfiler
+Distinct = profile.Distinct
+FeatureProfile = profile.FeatureProfile
+Histogram = profile.Histogram
+NumericStats = profile.NumericStats
+Quantiles = profile.Quantiles
+StringStats = profile.StringStats
+WordStats = profile.WordStats
+
+
+__all__ = [
+    "Distinct",
+    "Quantiles",
+    "Histogram",
+    "NumericStats",
+    "CharStats",
+    "WordStats",
+    "StringStats",
+    "FeatureProfile",
+    "DataProfile",
+    "DataProfiler",
+]
