+++ py-opsml/python/opsml/scouter/queue/__init__.py
@@ -1,0 +1,35 @@
+# type: ignore
+
+from .. import queue  # noqa: F401
+
+ScouterQueue = queue.ScouterQueue
+ScouterProducer = queue.ScouterProducer
+KafkaConfig = queue.KafkaConfig
+RabbitMQConfig = queue.RabbitMQConfig
+SpcServerRecord = queue.SpcServerRecord
+PsiServerRecord = queue.PsiServerRecord
+CustomMetricServerRecord = queue.CustomMetricServerRecord
+ServerRecord = queue.ServerRecord
+ServerRecords = queue.ServerRecords
+Feature = queue.Feature
+Features = queue.Features
+PsiFeatureQueue = queue.PsiFeatureQueue
+SpcFeatureQueue = queue.SpcFeatureQueue
+RecordType = queue.RecordType
+
+__all__ = [
+    "ScouterQueue",
+    "ScouterProducer",
+    "KafkaConfig",
+    "RabbitMQConfig",
+    "SpcServerRecord",
+    "PsiServerRecord",
+    "CustomMetricServerRecord",
+    "ServerRecord",
+    "ServerRecords",
+    "Feature",
+    "Features",
+    "PsiFeatureQueue",
+    "SpcFeatureQueue",
+    "RecordType",
+]
